Update README.md
Windows 10 has a limit of 25 items copied to the clipboard. Often, this is not sufficient, especially when programming. For that reason, ClipBoard Plus solves this issue by allowing you to have as many items in the clipboard as you need. Currently, it supports only text since that's what I use the most. However, if you would like multimedia elements to be supported, please let me know.